refactor(ui): move device status derivation into DeviceStatusOverview
DeviceStatusOverview already received the raw device object; deriving
rssi/battery color+icon and the model summary itself (instead of via
precomputed props from the parent) removes an inconsistent split
boundary flagged by the Standards-axis review.
