+++ packages/ui/src/components/device/DeviceStatusOverview.vue
@@ -1,26 +1,140 @@
 <script setup lang="ts">
 import type { Device } from '@/types'
-import { mdiAlert, mdiCheck, mdiContentCopy, mdiEye, mdiEyeOff } from '@mdi/js'
+import {
+  mdiAlert,
+  mdiBattery,
+  mdiBattery10,
+  mdiBattery20,
+  mdiBattery30,
+  mdiBattery40,
+  mdiBattery50,
+  mdiBattery60,
+  mdiBattery70,
+  mdiBattery80,
+  mdiBattery90,
+  mdiBatteryOutline,
+  mdiBatteryUnknown,
+  mdiCheck,
+  mdiContentCopy,
+  mdiEye,
+  mdiEyeOff,
+  mdiSignalCellular1,
+  mdiSignalCellular2,
+  mdiSignalCellular3,
+  mdiSignalCellularOutline,
+} from '@mdi/js'
 import { useClipboard } from '@vueuse/core'
-import { ref } from 'vue'
+import { computed, ref } from 'vue'
 import { VCol, VDivider, VIcon, VRow, VTextField, VTooltip } from 'vuetify/components'
+import { DEFAULT_RENDER_SIZE } from '@/utils/deviceRenderSize'
 import { formatDate } from '@/utils/formatDate'
 
-defineProps<{
+const props = defineProps<{
   device: Device
-  rssiColor: string
-  rssiIcon: string
-  batteryColor: string
-  batteryIcon: string
-  batteryPercentage: number
-  modelSummary: string
-  reportedSummary: string | null
-  reportMismatch: boolean
 }>()
 
 const { copy: copyToClipboard, copied: macCopied } = useClipboard()
 
 const showApikey = ref(false)
+
+const modelSummary = computed(() => {
+  const model = props.device.deviceModel
+  if (!model)
+    return `Not resolved yet — renders as TRMNL OG (${DEFAULT_RENDER_SIZE.width}x${DEFAULT_RENDER_SIZE.height})`
+  return `${model.label} (${model.width}x${model.height})`
+})
+
+const reportedSummary = computed(() => {
+  const { reportedModel, width, height } = props.device
+  const parts = [reportedModel, width && height ? `${width}x${height}` : null].filter(Boolean)
+  return parts.length > 0 ? parts.join(' · ') : null
+})
+
+const reportMismatch = computed(() => {
+  const { deviceModel, width, height } = props.device
+  if (!deviceModel || !width || !height)
+    return false
+  return width !== deviceModel.width || height !== deviceModel.height
+})
+
+const rssiStrength = computed(() => {
+  if (!props.device.rssi)
+    return -1
+  const rssi = Number.parseInt(props.device.rssi)
+  if (Number.isNaN(rssi))
+    return -1
+  if (rssi >= -70)
+    return 3
+  if (rssi >= -80)
+    return 2
+  if (rssi >= -90)
+    return 1
+  return 0
+})
+const rssiColor = computed((): string => {
+  switch (rssiStrength.value) {
+    case 3: return 'success'
+    case 2: return 'warning'
+    case 1: return 'warning'
+    case 0: return 'error'
+    case -1: return 'secondary'
+    default:
+      { const _: never = rssiStrength.value }
+      return ''
+  }
+})
+const rssiIcon = computed((): string => {
+  switch (rssiStrength.value) {
+    case 3: return mdiSignalCellular3
+    case 2: return mdiSignalCellular2
+    case 1: return mdiSignalCellular1
+    case 0: return mdiSignalCellularOutline
+    case -1: return mdiSignalCellularOutline
+    default:
+      { const _: never = rssiStrength.value }
+      return ''
+  }
+})
+
+const batteryPercentage = computed(() => {
+  if (!props.device.batteryVoltage)
+    return -1
+  const voltage = Number.parseFloat(props.device.batteryVoltage)
+  if (voltage >= 4.2)
+    return 100
+  if (voltage <= 3.0)
+    return 0
+  return Math.round(((voltage - 3.0) / 0.012))
+})
+
+const batteryColor = computed(() => {
+  if (batteryPercentage.value === -1)
+    return 'secondary'
+  if (batteryPercentage.value <= 10)
+    return 'error'
+  if (batteryPercentage.value <= 20)
+    return 'warning'
+  if (batteryPercentage.value <= 60)
+    return 'warning'
+  return 'success'
+})
+
+const batteryIcon = computed(() => {
+  const iconMap = [
+    { min: 95, icon: mdiBattery },
+    { min: 85, icon: mdiBattery90 },
+    { min: 75, icon: mdiBattery80 },
+    { min: 65, icon: mdiBattery70 },
+    { min: 55, icon: mdiBattery60 },
+    { min: 45, icon: mdiBattery50 },
+    { min: 35, icon: mdiBattery40 },
+    { min: 25, icon: mdiBattery30 },
+    { min: 15, icon: mdiBattery20 },
+    { min: 5, icon: mdiBattery10 },
+    { min: 0, icon: mdiBatteryOutline },
+  ]
+  return iconMap.find(item => batteryPercentage.value >= item.min)?.icon || mdiBatteryUnknown
+})
 </script>
 
 <template>
